extract attachments casl ability in separate file
