+++ src/casl/casl-ability.factory.ts
@@ -10,7 +10,6 @@ import { ConfigService } from "@nestjs/config";
 import { JobConfigService } from "src/config/job-config/jobconfig.service";
 import { JWTUser } from "src/auth/interfaces/jwt-user.interface";
 import { AccessGroupsType } from "src/config/configuration";
-import { Attachment } from "src/attachments/schemas/attachment.schema";
 import { Datablock } from "src/datablocks/schemas/datablock.schema";
 import { DatasetClass } from "src/datasets/schemas/dataset.schema";
 import { Instrument } from "src/instruments/schemas/instrument.schema";
@@ -29,6 +28,7 @@ import { SampleClass } from "src/samples/schemas/sample.schema";
 import { User } from "src/users/schemas/user.schema";
 import { Action } from "./action.enum";
 import { Subjects, PossibleAbilities, Conditions } from "./types/casl-subjects";
+import { AttachmentAbility } from "./abilities/attachments.ability";
 
 export type AppAbility = MongoAbility<PossibleAbilities, Conditions>;
 
@@ -37,6 +37,7 @@ export class CaslAbilityFactory {
   constructor(
     private configService: ConfigService,
     private jobConfigService: JobConfigService,
+    private attachmentAbility: AttachmentAbility,
   ) {
     this.accessGroups =
       this.configService.get<AccessGroupsType>("accessGroups");
@@ -75,83 +76,7 @@ export class CaslAbilityFactory {
   }
 
   attachmentAccess(user: JWTUser) {
-    const { can, build } = new AbilityBuilder(
-      createMongoAbility<PossibleAbilities, Conditions>,
-    );
-    const ifPublished = { isPublished: true };
-
-    /**
-     * Unauthenticated user
-     */
-    can(Action.AttachmentRead, Attachment, ifPublished);
-
-    if (!user) {
-      return build({
-        detectSubjectType: (item) =>
-          item.constructor as ExtractSubjectType<Subjects>,
-      });
-    }
-
-    const ifOwner = { ownerGroup: { $in: user.currentGroups } };
-    const ifAccess = { accessGroups: { $in: user.currentGroups } };
-
-    /**
-     * Authenticated user
-     */
-    can(Action.AttachmentRead, Attachment, ifOwner);
-    can(Action.AttachmentRead, Attachment, ifAccess);
-    can(Action.AttachmentRead, Attachment, ifPublished);
-
-    if (
-      user.currentGroups.some((g) =>
-        this.accessGroups?.attachment.includes(g),
-      ) ||
-      this.accessGroups?.attachment.includes("#all")
-    ) {
-      /**
-       * User belonging to ATTACHMENT_GROUPS
-       */
-      can(Action.AttachmentCreate, Attachment, ifOwner);
-      can(Action.AttachmentUpdate, Attachment, ifOwner);
-      can(Action.AttachmentDelete, Attachment, ifOwner);
-    }
-
-    if (
-      user.currentGroups.some((g) =>
-        this.accessGroups?.attachmentPrivileged.includes(g),
-      )
-    ) {
-      /**
-       * User belonging to ATTACHMENT_PRIVILEGED_GROUPS
-       */
-      can(Action.AttachmentCreate, Attachment);
-      can(Action.AttachmentUpdate, Attachment, ifOwner);
-      can(Action.AttachmentDelete, Attachment, ifOwner);
-    }
-
-    if (user.currentGroups.some((g) => this.accessGroups?.admin.includes(g))) {
-      /**
-       * User belonging to ADMIN_GROUPS
-       */
-      can(Action.AccessAny, Attachment);
-
-      can(Action.AttachmentCreate, Attachment);
-      can(Action.AttachmentRead, Attachment);
-      can(Action.AttachmentUpdate, Attachment);
-      can(Action.AttachmentDelete, Attachment);
-    }
-
-    if (user.currentGroups.some((g) => this.accessGroups?.delete.includes(g))) {
-      /**
-       * User belonging to DELETE_GROUPS
-       */
-      can(Action.AttachmentDelete, Attachment);
-    }
-
-    return build({
-      detectSubjectType: (item) =>
-        item.constructor as ExtractSubjectType<Subjects>,
-    });
+    return this.attachmentAbility.buildAbility(user);
   }
 
   datasetEndpointAccess(user: JWTUser) {
