+++ src/casl/casl.module.ts
@@ -2,9 +2,11 @@ import { Module } from "@nestjs/common";
 import { ConfigModule } from "@nestjs/config";
 import { CaslAbilityFactory } from "./casl-ability.factory";
 import { JobConfigModule } from "src/config/job-config/jobconfig.module";
+import { AttachmentAbility } from "./abilities/attachments.ability";
+
 @Module({
   imports: [JobConfigModule, ConfigModule],
-  providers: [CaslAbilityFactory],
+  providers: [CaslAbilityFactory, AttachmentAbility],
   exports: [CaslAbilityFactory],
 })
 export class CaslModule {}
